+++ packages/components-providers/src/index.ts
@@ -24,4 +24,7 @@
 
  */
 export * from './ComponentsProvider'
+export * from './ExtendComponentsProvider'
+export * from './FocusTrap'
 export * from './ScrollLock'
+export * from './TrapStack'