+++ packages/components/src/Form/Inputs/Combobox/utils/useInputEvents.ts
@@ -25,7 +25,6 @@
  */
 
 import {
-  FocusEvent,
   MouseEvent as ReactMouseEvent,
   useRef,
   useContext,
@@ -96,13 +95,8 @@ export function useInputEvents<
 
   const handleBlur = useBlur(context)
 
-  function handleFocus(e: FocusEvent<HTMLInputElement>) {
-    const input = e.currentTarget
-    if (readOnly && input) {
-      window.requestAnimationFrame(() => {
-        input.selectionEnd = input.selectionStart = 0
-      })
-    } else if (selectOnClick) {
+  function handleFocus() {
+    if (!readOnly && selectOnClick) {
       selectOnClickRef.current = true
     }