+++ packages/components/src/Text/CodeBlock.tsx
@@ -46,7 +46,7 @@ const CodeBlockLayout: FC<CodeBlockProps> = ({
 export const CodeBlock = styled(CodeBlockLayout)`
   ${border}
   color: ${({ theme }) => theme.colors.text};
-  overflow-y: scroll;
+  overflow-y: auto;
 
   code {
     font-family: inherit;