+++ CHANGELOG.md
@@ -14,6 +14,20 @@ NOTE: This is the CHANGELOG for the @looker/components MONOREPO. Each package ha
 - [Design Tokens](./packages/design-tokens/CHANGELOG.md)
 - [Icons](./packages/icons/CHANGELOG.md)
 
+## [0.9.25]
+
+### Changed
+
+Library updated to use Styled Components 5 & Typescript 3.9.x
+
+### Added
+
+- update InputDate and InputDateRange to support disabled and readOnly
+
+### Fixed
+
+- Update DataTable to support onClick of `Link/Anchor` as text inside rows.
+
 ## [0.9.23]
 
 ### Added